Move FactoryQuerier to mock_querier
Integrate with WasmMockQuerier
